Add stats dashboard and UI improvements

+++ static/app.js
@@ -35,7 +35,8 @@ document.addEventListener("DOMContentLoaded", function() {
 		
 		    document.getElementById("textToAnalyze").value = "";
 		
-		    await loadHistory();
+		    await refreshUI();
+			
 		} finally {
 			button.disabled = false;
 			button.textContent = "SUBMIT";
@@ -45,13 +46,21 @@ document.addEventListener("DOMContentLoaded", function() {
 	document.getElementById("emotionFilter").addEventListener("change", loadHistory)
 	document.getElementById("sentimentFilter").addEventListener("change", loadHistory)
 	
-    loadHistory();
+    refreshUI();
 	
 	document.getElementById("saveEditBtn").addEventListener("click", async function () {
 	    const text = document.getElementById("updateText").value;
 	
 	    await updateEntry(currentEditId, text);
     });
+	
+	const editScreen = document.getElementById("editScreen");
+	
+	editScreen.addEventListener("shown.bs.modal", function () {
+		const input = document.getElementById("updateText");
+		input.focus();
+		input.select();
+	});
 	
 });
 
@@ -62,6 +71,18 @@ async function loadHistory() {
 	const data = await response.json();
 	
 	renderCards(data);
+}
+
+async function loadStats() {
+	const response = await fetch("/stats/");
+	const stats = await response.json();
+	
+	renderStats(stats);
+}
+
+async function refreshUI() {
+	await loadHistory();
+	await loadStats();
 }
 
 function renderCards(data) {
@@ -99,44 +120,108 @@ function renderCards(data) {
 		    });
 		}
 		
+		const editedHtml = updatedTime
+		    ? `• Edited: ${updatedTime}`
+			: "";
+		
 	    historyDiv.innerHTML += `
-            <div class="card mb-3">
+            <div class="card mb-3 shadow-sm">
                 <div class="card-body">
     	            <div class="row">
 			            <div class="col-sm-2">
-				            <h3><span class="badge ${emotionClass} me-2">${entry.emotion}</span></h3>
+				            <h3><span class="badge ${emotionClass} me-2" style="min-width: 8vw">${entry.emotion}</span></h3>
 				        </div>
 				        <div class="col-sm-2">
-				            <h3><span class="badge ${sentimentClass}">${entry.sentiment}</span></h3>
+				            <h3><span class="badge ${sentimentClass}" style="min-width: 8vw">${entry.sentiment}</span></h3>
 				        </div>
-				        <div class="col-sm-5"></div>
-						<div class="col-sm-1">
-						    <button
-							    class="btn btn-warning"
-								onclick="openEditModal(${entry.id}, \`${entry.text}\`)"
-								data-bs-toggle="modal"
-								data-bs-target="#editScreen">
-								EDIT
-							</button>
+				        <div class="col-sm-8"></div>
+					</div>
+					<div class="d-flex align-items-center justify-content-end h-100">
+					    <div class="p-2 flex-grow-1">
+					        <p class="card-text">${entry.text}</p>
 						</div>
-				        <div class="col-sm-2">
-				            <button onclick="deleteEntry(${entry.id})" type="button" class="btn btn-danger">DELETE</button>
-				        </div>
+						<button
+						    class="btn btn-warning mb-3"
+							onclick="openEditModal(${entry.id}, \`${entry.text}\`)"
+							data-bs-toggle="modal"
+							data-bs-target="#editScreen"
+							style="min-width: 6vw">
+							EDIT
+						</button>
+						<div class="p-2"></div>
+				        <button onclick="deleteEntry(${entry.id})" type="button" class="btn btn-danger mb-3" style="min-width: 6vw">DELETE</button>
 			        </div>
-                    <p class="card-text">${entry.text}</p>
-		    	    <small class="text-muted">${localTime}</small>
-		`
-		
-		if (updatedTime) {
-		    historyDiv.innerHTML += `<small class="text-muted">Edited at: ${updatedTime}</small>`;
-	    }
-		
-		historyDiv.innerHTML += `
+		    	    <small><small class="text-muted d-block mb-2">Created: ${localTime} ${editedHtml}</small></small>
                 </div>
             </div>
         `;
 		
 	});
+}
+
+function formatEmotion(e) {
+	return e.charAt(0).toUpperCase() + e.slice(1);
+}
+
+async function renderStats(stats) {
+	
+	document.getElementById("stats").innerHTML = `
+	    <div class="row mb-4">
+		    
+			<div class="col-md-3">
+			    <div class="card shadow-sm text-center h-100">
+				    <div class="card-body">
+					    <h6 class="text-muted">Total Entries</h6>
+						<h3>${stats.total}</h3>
+					</div>
+				</div>
+			</div>
+			
+			<div class="col-md-3">
+			    <div class="card shadow-sm text-center h-100">
+				    <div class="card-body">
+					    <h6 class="text-muted">Positive</h6>
+						<h3 class="text-success">${stats.positive}</h3>
+					</div>
+				</div>
+			</div>
+			
+			<div class="col-md-3">
+			    <div class="card shadow-sm text-center h-100">
+				    <div class="card-body">
+					    <h6 class="text-muted">Negative</h6>
+						<h3 class="text-danger">${stats.negative}</h3>
+					</div>
+				</div>
+			</div>
+			
+			<div class="col-md-3">
+			    <div class="card shadow-sm text-center h-100">
+				    <div class="card-body">
+					    <h6 class="text-muted">Neutral</h6>
+						<h3 class="text-primary">${stats.neutral}</h3>
+					</div>
+				</div>
+			</div>
+			
+			<div class="col-12">
+			    <div class="card shadow-sm text-center h-100">
+				    <div class="card-body">
+					    <h6 class="text-muted">Top Emotion</h6>
+						<h2 class="text-success mt-3">${stats.top_emotions.length > 1
+						                            ? "Tie: "
+						                            : ""} 
+						                        ${stats.top_emotions.length
+						                            ? stats.top_emotions
+												       .map(formatEmotion)
+													   .join(" • ")
+													: "-"}</h2>
+					</div>
+				</div>
+			</div>
+			
+		</div>
+	`;
 }
 
 function getEmotionClass(emotion) {
@@ -211,7 +296,9 @@ function openEditModal(id, text) {
 	document.getElementById("saveEditBtn").textContent = "Update";
 	
 	currentEditId = id;
+	
 	document.getElementById("updateText").value = text;
+	
 }
 
 function closeEditModal() {
@@ -221,15 +308,26 @@ function closeEditModal() {
 	    new bootstrap.Modal(modalElement);
 		
 	modal.hide();
+	
+	document.getElementById("updateText").value = "";
+    document.getElementById("editResult").innerHTML = "";
+    currentEditId = null;
+		
 }
 
 async function deleteEntry(id) {
+	
+	const confirmed = confirm("Are you sure you want to delete this entry?");
+	
+	if (!confirmed) {
+		return;
+	}
 			
 	await fetch(`/entries/${id}`, {
 	    method: "DELETE"
 	});
 			
-	await loadHistory();
+	await refreshUI();
 }
 
 async function updateEntry(id, text) {
@@ -253,12 +351,7 @@ async function updateEntry(id, text) {
 	    });
 		
 		closeEditModal();
-		await loadHistory();
-		
-		document.getElementById("updateText").value = "";
-        document.getElementById("editResult").innerHTML = "";
-        currentEditId = null;
-		
+		await refreshUI();
 	}
 	finally {
 		button.disabled = false;
